Add autocomplete progressive enhancement js
