+++ common/static/common/js/autocomplete.js
@@ -1,27 +1,12 @@
 import accessibleAutocomplete from 'accessible-autocomplete'
+import { removeNewLine, cleanResults } from './util.js'
 
 // Only a new result has a label, else result is the original result's label
 const template = (result) => typeof result == "object" ? result.label : result
 
 let aborter = null;
 
-const newLine  = /\n/g;
-const removeNewLine = (str) => str.replace(newLine, "")
-
-const cleanResults = (results) => {
-  /*
-  Results which contain new line characters are considered as a new query when
-  selected, causing results to be repopulated unnecessarily. To prevent this,
-  we can remove new line characters from results since they're unimportant here.
-  */
-  for(let i = 0, len = results.length; i < len; i++) {
-    results[i].label = removeNewLine(results[i].label);
-  }
-  return results
-}
-
-
-const autoCompleteElement = (element, includeNameAttr=true) => {
+const autoCompleteElement = (element, includeNameAttr = true) => {
   const hiddenInput = element.querySelector("input[type=hidden]");
   const params = {
     element: element.querySelector(".autocomplete-container"),
@@ -31,12 +16,12 @@ const autoCompleteElement = (element, includeNameAttr=true) => {
       const searchParams = new URLSearchParams();
       searchParams.set("search", query);
       searchParams.set("format", "json");
-      if(aborter) {
+      if (aborter) {
         aborter.abort();
       }
       aborter = new AbortController();
       const signal = aborter.signal
-      fetch(`${source_url}?${searchParams}`, {signal})
+      fetch(`${source_url}?${searchParams}`, { signal })
         .then(response => response.json())
         .then(data => populateResults(cleanResults(data.results)))
         .catch(err => console.log(err));
@@ -67,7 +52,7 @@ const autoCompleteElement = (element, includeNameAttr=true) => {
 }
 
 
-const initAutocomplete = (includeNameAttr=true) => { 
+const initAutocomplete = (includeNameAttr = true) => {
   for (let element of document.querySelectorAll(".autocomplete")) {
     autoCompleteElement(element, includeNameAttr);
   }
